jira-mcp/requirements.txt
langchain-mcp-adapters
langgraph
langchain-openai
langchain
python-dotenv
mcp
streamlit
fastapi
uvicorn